+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Method `getSelectedVariable` in
+  [Container](https://chird.github.io/meteoJS/doc/module-meteoJS_modelviewer_container.Container.html).
 
 ## [1.12.0] - 2021-03-01
 ### Added